refactor: msrpScraper to msrp_scraper to match Python naming convention
