+++ src/components/utilities/scraperApi.py
@@ -1,5 +1,5 @@
 from flask import Flask, request, jsonify
-from msrpScraper import get_car_data
+from msrp_scraper import get_car_data
 from flask_cors import CORS
 from concurrent.futures import ThreadPoolExecutor
